Split GiniScreenAPICoordinator into several files.
